Fix uci packets.
We could not parse the HAL's response to the first message we sent it.
Looking through the specification, it seems like we were not ordering
some of our fields correctly: the PDL states that fields should be
ordered from least to most significant bit.  This reorders the fields
and fixes all of the relevant tests.

Test: libuwb_uci_packet_tests
Test: Send specific packet to the HAL and parse its response.
Change-Id: I20a168f40a376f60c39a2980673846f26f7c176e