Add an API for blocking HAL commands
This adds an API to the dispatcher that will block until a response is
received.

It currently only handles a single such call being outstanding at a
time.  If we need to extend this in the future, we can simply use a
map of channels keyed on whatever differentiates them.

Note that we could also do a similar approach for communicating
between the Dispatcher and the HAL.  That would allow us to replace
the current asynchronous architecture (where the Dispatcher sends a
message to the HAL, returns to its select loop, and later receives a
response) with a similar blocking API, making communication to the HAL
into straightline code.  This would have the added benefit of making
it easier to handle multiple outstanding requests (see the previous
paragraph), as the work for doing so would be pushed down to the
adaptation.

Bug: 197341298
Test: libuwb_uci_rust_tests
Test: Call with actual real data and see a real result sent back.
Change-Id: I8d8475f980483ac9ea3d4ce43bbbbb72c1e918df