Add code to call back to the JNI.
This adds some infrastructure for calling back into the JNI.  It
implements three of the five callbacks we need.

Note that it cannot yet be called as we have not yet initialized the
Java class to handle initializations.

As the long TODO suggests, we should potentially reconsider how we
attach to the thread once we have a better sense of how this is used
and what its performance is.

Bug: 197341298
Test: Manually insert code that calls it.
Test: Manually modify Java callbacks to throw an Exception.
Change-Id: I72f231096e44e2f1db33cf6f467ecb662f986d3d