Allow the adaptation's callback to have state.
The current implementation does not allow the callbacks to have any
state, which they need so they can communicate to the dispatcher.
This fixes that, and additionally gives them a channel that they will
likely use to communicate.  The exact API might need to change when it
is fully implemented, but this should show how that can be done.

Bug: 197341298
Test: libuwb_uci_rust_tests
Change-Id: I3a3929f6dda106caeb9d53e0dbdf7746af5f2ad7