Parse HAL responses and send them to the dispatcher.
This hooks up the adaptation so that it can send responses from the
HAL to the dispatcher.

The main part of the commit is modifying our parsing so that it
returns its result rather than sending it to a callback (which is
mainly so that we don't have to pass around the state that callback
would need).  In the process, we also add some new error types and
improve our test.

Bug: 197341298
Test: libuwb_uci_rust_tests
Change-Id: Ibecadba16a49b74f9f9b2fbaf9ddea261a4a6820