uwb(uci-rust): Don't crash on failure to convert to i32 session_id
session_id is defined in the spec as 4 byte unsigned value. The HAL
interface only allows signed values, so we need to just cast the
received 4 byte value received via the UCI packet as i32 for the HAL call.

The alternative is to declare session_id as long in the HAL interface
and upconvert from u32 to i64.

Bug: 216104681
Test: Compiles
Change-Id: I050c1985b680246726c3a1328755aba487b72390