uwb(uci-pdl): Fix UCI packet heirarchy (use scalars for opcode)
This uses the raw values for the opcodes to workaround the limitation
of being unable to set enum values in scalar fields. This helps fix some
of the redundancies in the Vendor Cmd/Rsp/Ntf packets.

Also, reordered the packets around to ensure they are in the correct
order (according to opcode).

Bug: b/202760099
Test: Unit tests
Change-Id: Ib8f32a229528c80b8b7c39f4b866b4a9605a5e2e