uci(rust): Handle fragmentation/defragmentation of packets
UCI spec mandates fragmentation of cmd/rsp/ntf if the
payload size if > 255 (len field is 8 byte).
Ideally this should be handled by the PDL more cleanly in the future
But, we need a fix for UWB in Android T.

Changes:
1. Add a new base |UciPacket| definition which can hold the whole
cmd/rsp/notif and is used throughout the UWB stack. Use this as the base
packet for all packet definitions in the PDL. This allows the stack
to abstract the fragmentation to |UwbAdaption| module.
2. Add a separate |UciPacketHal| definition representing the spec defined UCI
packet. This packet is sent/received from the HAL interface.
3. Add conversion mechanism from |UciPacket| to |UciPacketHal| &
vice-versa for incoming/outgoing HAL packets.
4. added a MockHal for testing of UwbAdaptationImpl.

Bug: 202760099
Test: Existing Manual tests.
Test: atest CtsUwbTestCases
Test: atest libuwb_uci_rust_tests
Change-Id: Id275000cf27e10d6e35285c5bfb2beca095f6d75