uwb_core: make FiraAppConfigParams generate tlvs for every field
At ag/17750782, FiraAppConfigParams supports to generate the updated
tlvs from the difference of the old parameters. To avoid the
difference algorithm doesn't handle the situation that a key was set
before but unset after, this CL makes FiraAppConfigParams generate
tlvs for every field. With this CL, the set of keys are always the
same, so we won't encounter the situation.

Bug: 228136039
Test: cargo test
Change-Id: I623adf61ed5866fe1cd2e6aedd07fc5fdde6b0a8