uwb_core: reset the service when UCI is in error state
The UCI HAL sends a notification when it enters a error state. Also,
the UCI HAL might be stuck at a weird state when the expected response
or notification is not received in timeout.

This CL forces resetting the UCI HAL and clears the internal state in
these situations, and hopes the HAL goes back to the normal state.

Bug: 228395837
Test: cargo test
Test: cargo clippy
Change-Id: I778f3abc96df2cd8141338b69d6b2bfa687ad088