uwb_core: unify and export params module
Originally, both the "uci" and "session" modules have a "params"
sub-module. The components inside the params sub-modules should be
exposed to the caller of the library. However, the other part of the
"session" module doesn't need to be public.

This CL creates a new public "params" module, and moves the original
sub-modules into this new module. Also, the params module re-exposes
all the components from its sub-modules. The caller could access the
components simply by "uwb_core::params::<component>".

Bug: 228395837
Test: cargo test
Test: cargo clippy

Change-Id: I1ce81cf88e180d52b10837ff5f23ba18ef961526