uwb(rust): Handle HAL close failure when HAL has crashed
As a part of the stack recovery, we toggle off UWB & then start it
again. This fails currently because the native stack does not exit as
expected on doDenitialize(). This was because the HAL close API fails
(because the HAL is dead) and we would not receive the HAL close
completed callback. So, instead exit the native stack immediately if the
HAL close invocation fails for some reason.

Bug: 236392438
Test: Manual steps:
1. Start a ranging session.
2. Kill vendor HAL.
3. Wait for recovery.
4. Start a new ranging session and ensure that it succeeds.

Change-Id: I709ead843a9632147a131b9ba80c08b0f4658624