uwb_core: Add UciManagerSync interface
This CL adds a sync interface at Uci Command level. The existing Android
uwb_uci stack in particular, and most JNI in general, expects a
synchornized interface, whereas UciManager trait is  async.
UciManagerSync is introduced to bridge the gap. NotificationManager,
similar to EventManager in previous stack is introduced, and a system
dependent implementation is expected.

Test: cargo clippy
Test: cargo test
Test: mmm .
Bug: 235453225

Change-Id: I784b4a0fab9e8caee6883f6dea88d04808bbaa4e