uwb_core: split UciCommand to multiple packet
At CL:17531741, we introduced the mechanism to fragment a UCI message
to a list of UCI packets. The CL:17615424 applied the fragmentation
to uwb_core library for the notifications and response, but forgot the
command fragmentation.

This CL fragment the UCI command to the list of UCI packets. Besides,
it renames RawUciMessage to UciHalPacket to fit the meaning of the
type properly.

Bug: 202760099
Test: cargo test
Change-Id: Id8d0a675040d634bf5a21f5654bdf8e88c430461