Add message control and subsession key to controlee.
Message control and subsession key are fields required when Provisioned
STS is used. But, we cannot directly add these fields because of
backward compatability. So we come up with this workaround to fix the
compatability problem: when using provisioned STS, we will switch the
command from SessionUpdateControllerMulticastList to
SessionUpdateControllerMulticastListV2.

Test: atest ServiceUwbTests, libuwb_uci_rust_tests,
libuwb_uci_packet_tests
Bug: 232453347
Bug: 231581242

Change-Id: Id2bded30b073eea46c6cb803c693806dea3823a3