uwb_core: make the Runtime outlives UwbService
The tokio Runtime has a implicit rule: one Runtime could not be
dropped in the asynchronous context of another Runtime. To prevent
this situation, the better practice is to only create one Runtime as
the "environment" that outlives all the objects that needs
asynchronous context.

Originally, the UwbService owns the working tokio Runtime. This CL
changes to create the Runtime outside the UwbService, and only injects
the handle of the Runtime into UwbService.

Bug: 256721796
Test: cargo test
Change-Id: I1fdd627fb186233f625367d1fd7f591654480495