uwb_core: Move no-op UciHal and UwbServiceCallback into library
Originally, the library doesn't provide the no-op UciHal and
UwbServiceCallback implementation. All the users of the UwbService
should implement the traits. But actually the no-op implementation is
useful for testing.

This CL moves the no-op implementation from examples/main.rs into the
library. Also, it renames UciLoggerNull and UciLoggerFactoryNull to
NopUciLogger and NopUciLoggerFactory for consistency.

Bug: 259205925
Test: cargo test
Test: mmm packages/modules/Uwb

Change-Id: Ieff0e3c25323e8987b1b4473f33a17c76e8ead31