Data Packet Send changes.
Fix issue where the UciManagerActor thread is stuck after sending a Data
packet fragment and unable to process any Rx packets (like
DATA_CREDIT_NTF and DATA_TRANSFER_STATUS_NTF).

The solution is to move the fragmentation into the UciManager, send each
data packet fragment to the UciManagerActor, and move the wait for the
corresponding NTF packets to the UciManager.

Bug: 266016401
Test: atest libuwb_core_test libuwb_uci_packet_tests libuwb_uci_jni_rust_tests
Test: Data Packet Tx/Rx tested
Change-Id: Ia0a00a9bdebcb2f17f8cf155ae79caac03e6cd86