DO NOT MERGE: Use protobuf v2.x
The protobuf crate is being updated to version 3.x. Keep
current users of the protobuf crate on 2.x for now.

This is functionally a no-op, but cannot be merged into
tm-mainline-prod due to missing dependencies, thus the
"DO NOT MERGE."

Test: Treehugger
Bug: 270895633
Merged-in: I92197ea6b91da19b32074b981c5e6f02cb32d82e
Change-Id: I8c9114fbbf6827877a1059054497078812596203