Fragment UCI DATA SND packets based upon UWBS capabilities.
Currently the data packets sent by Host to UWBS are fragmented with a
fixed payload size of 255 bytes.

This CL changes the behavior to use the UWBS capability value "MAX_DATA_PACKET_PAYLOAD_SIZE" to determine
the max payload size of each data packet fragment. It will still use the
value 255 as the default (in case the capability couldn't be read from
the UWBS).

Test: atest libuwb_uci_jni_rust_tests libuwb_core_tests libuwb_uci_packet_tests
Bug: 301834443
Change-Id: I36866609ac8cbc091d66a134a17f923d7d330232