Make libuci_hal_android_tests available to the platform
libuci_hal_android_tests was hitting a quirk in Soong where a module
that was not available for the platform but had no apexes that depended
on it was treated as a platform variant.  That quirk is being fixed as
part of preparing the apex info mutator for incremental analysis.
Explicitly mark libuci_hal_android_tests as available to the platform.

Bug: 372543712
Test: builds
Change-Id: I6b5ce52c978afb03c2a4fc05bd09f9ebbe6e73e8